+++ modules/Trades/TradingJournal.Modules.Trades/ViewModel/TradingStatisticViewModel.cs
@@ -2,9 +2,9 @@ namespace TradingJournal.Modules.Trades.ViewModel;
 
 public sealed record TradingStatisticViewModel
 {
-    public double TotalPnL { get; set; }
+    public decimal TotalPnL { get; set; }
 
-    public double WinRate { get; set; }
+    public decimal WinRate { get; set; }
 
     public int TotalTrades { get; set; }
